Add progress as of 2019-06-04
This isn't a great commit, but I realised I had loads of uncommited work and I
got anxious about loosing it. Loads of stuff in here, I think some styling
of <p>s to avoid as many <br>s as I had, along with a bunch of work on the
infinity trees where and why I was wrong article.